+++ src/services/index.ts
@@ -1,7 +1,6 @@
 export * from './users.service';
 export * from './projects.service'
 export * from './builds.service'
-export * from './tests.service'
 export * from './static.service'
 export * from './testVariation.service'
 export * from './testRun.service'
